+++ multibase/src/commonMain/kotlin/at/asitplus/io/UVarInt.kt
@@ -8,15 +8,16 @@ package at.asitplus.io
 /**
  * Unsigned variable-length integer supporting values up to 2^63 - 1.
  */
-class UVarInt private constructor(private val number: ULong) {
+data class UVarInt(private val number: ULong) {
+    init { require(number <= Long.MAX_VALUE.toULong()) }
     /**
      * Convenience constructor to create an object from an unsigned int. To create larger UVarInts, use [fromByteArray].
      */
     constructor(number: UInt) : this(number.toULong())
 
-    /**
-     * Returns the ULong value of this UVarInt.
-     */
+    override fun toString() = "0x"+number.toString(16)
+
+    /** Returns the ULong value of this UVarInt. */
     fun toULong(): ULong = number
 
     /**
@@ -25,13 +26,16 @@ class UVarInt private constructor(private val number: ULong) {
     fun encodeToByteArray(): ByteArray {
         var acc = number
         var i = 0
-        val res = mutableListOf<Byte>()
+        // we never encode anything larger than 2^63-1 (which fits in 9*7 bits)
+        var res = ByteArray(MAX_BYTES)
         while (acc >= 0x80u) {
-            res += (((acc and 0x7Fu) or 0x80u).toByte())
+            res[i++] = (((acc and 0x7Fu) or 0x80u).toByte())
             acc = (acc.toLong() ushr 7).toULong()
-            i++
+            require(i < MAX_BYTES)
         }
-        return (res + (acc and 0x7Fu).toByte()).toByteArray()
+        res[i++] = (acc and 0x7Fu).toByte()
+        if (i == MAX_BYTES) return res
+        else return res.copyOf(i)
     }
 
     companion object {
@@ -39,22 +43,57 @@ class UVarInt private constructor(private val number: ULong) {
          * Maximum number of bytes representing a UVarInt in this encoding,
          * supporting values up to 2^63 - 1.
          */
-        const val MAX_BYTES = 9L
+        const val MAX_BYTES = 9
 
         /**
          * Decodes a varint-encoded ByteArray into a UVarInt.
          * @throws NumberFormatException on illegal input (e.g. values larger than 2^63 - 1 or non-minimal encodings)
          */
         @Throws(NumberFormatException::class)
-        fun fromByteArray(bytes: ByteArray): UVarInt = UVarInt(decode(bytes))
+        @Deprecated(message="Specify explicitly whether you want to consume the entire byte array" +
+                "(use fromByteArrayStrict or fromByteArrayPermissive)",
+            replaceWith = ReplaceWith("fromByteArrayStrict(bytes)"))
+        fun fromByteArray(bytes: ByteArray): UVarInt = fromByteArrayPermissive(bytes).first
 
-        private fun decode(encoded: ByteArray): ULong {
+        /**
+         * Decodes the entirety of [bytes], starting at index [startIndex], into an [UVarInt].
+         * @throws NumberFormatException on illegal input (e.g. values larger than 2^63 - 1 or non-minimal encodings)
+         * @throws NumberFormatException if [bytes] contains trailing bytes after the [UVarInt]
+         * @see fromByteArrayPermissive
+         */
+        @Throws(NumberFormatException::class)
+        fun fromByteArrayStrict(bytes: ByteArray, startIndex: Int = 0): UVarInt =
+            decode(bytes, startIndex).let {
+                if(it.second != bytes.size) {
+                    throw NumberFormatException(
+                        "UVarInt decoding only consumed ${it.second} bytes, but ${bytes.size} were provided.")
+                }
+                UVarInt(it.first)
+            }
+
+        /**
+         * Decodes the leading UVarInt from [bytes], starting at index [startIndex].
+         * Returns the UVarInt, and the first index in [bytes] past the UVarInt.
+         * @throws NumberFormatException on illegal input (e.g. values larger than 2^63 - 1 or non-minimal encodings)
+         * @see fromByteArrayStrict
+         */
+        @Throws(NumberFormatException::class)
+        fun fromByteArrayPermissive(bytes: ByteArray, startIndex: Int = 0): Pair<UVarInt,Int> =
+            decode(bytes, startIndex).let { Pair(UVarInt(it.first),it.second) }
+
+        /** Decodes a varint-encoded leading ByteArray */
+        private fun decode(encoded: ByteArray, startIndex: Int): Pair<ULong,Int> {
+            require(startIndex < (Int.MAX_VALUE - MAX_BYTES))
             var value = 0uL
+            var i = startIndex
             var s = 0
-            var i = 0
+            val last = startIndex + MAX_BYTES
             while (true) {
-                val uByte = encoded[i].toUByte()
-                if ((i == 8 && uByte >= 0x80u) || i >= MAX_BYTES) {
+                if (i >= encoded.size) {
+                    throw NumberFormatException("varint is not terminated (overruns array)")
+                }
+                val uByte = encoded[i++].toUByte()
+                if ((i == last && uByte >= 0x80u) || i > last) {
                     // this is the 9th and last byte we're willing to read, but it
                     // signals there's more (1 in MSB).
                     // or this is the >= 10th byte, and for some reason we're still here.
@@ -64,11 +103,10 @@ class UVarInt private constructor(private val number: ULong) {
                     if (uByte == 0u.toUByte() && s > 0) {
                         throw NumberFormatException("varint not minimally encoded")
                     }
-                    return value or (uByte.toULong() shl s)
+                    return Pair(value or (uByte.toULong() shl s), i)
                 }
                 value = value or ((uByte and 0x7fu).toULong() shl s)
                 s += 7
-                i++
             }
         }
     }
@@ -79,4 +117,16 @@ class UVarInt private constructor(private val number: ULong) {
  * @throws NumberFormatException on illegal input (e.g. values larger than 2^63 - 1 or non-minimal encodings)
  */
 @Throws(NumberFormatException::class)
+@Deprecated("Explicitly specify whether you want a leading UVarInt or the full byte array as UVarInt",
+    ReplaceWith("decodeAsUVarInt()"))
 fun ByteArray.varIntDecode() = UVarInt.fromByteArray(this)
+
+@Throws(NumberFormatException::class)
+/** @see UVarInt.fromByteArrayStrict */
+fun ByteArray.decodeAsUVarInt(startIndex: Int = 0) =
+    UVarInt.fromByteArrayStrict(this, startIndex)
+
+@Throws(NumberFormatException::class)
+/** @see UVarInt.fromByteArrayPermissive */
+fun ByteArray.decodeLeadingUVarInt(startIndex: Int = 0) =
+    UVarInt.fromByteArrayPermissive(this, startIndex)